+++ KaguyaProjectV2/KaguyaBot/DataStorage/DbData/Models/FavoriteTrack.cs
@@ -10,5 +10,11 @@ namespace KaguyaProjectV2.KaguyaBot.DataStorage.DbData.Models
         public ulong UserId { get; set; }
         [Column(Name = "track_id"), NotNull]
         public string TrackId { get; set; }
+        [Column(Name = "track_title"), NotNull]
+        public string TrackTitle { get; set; }
+        [Column(Name = "track_author"), NotNull]
+        public string TrackAuthor { get; set; }
+        [Column(Name = "track_duration"), NotNull]
+        public double TrackDuration { get; set; }
     }
 }